+++ GestionEscolar/DTO/PonderacionesResponse.cs
@@ -8,6 +8,6 @@
     public class PonderacionResponse
     {
         public string nombre { get; set; }
-        public int ponderacion { get; set; }
+        public float ponderacion { get; set; }
     }
 }
